dbms project/requirements.txt
streamlit
mysql.connector